Fetch & sync trakt followed shows.

+++ shared/domain/trakt/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/trakt/api/model/TraktCreateListResponse.kt
@@ -14,6 +14,6 @@ data class TraktCreateListResponse(
 
 @Serializable
 data class ListIds(
-    @SerialName("trakt") val trakt: Long,
+    @SerialName("trakt") val trakt: Int,
     @SerialName("slug") val slug: String
 )
